Create better policy rules
Prior to this commit the policy was targeting all types of Kubernetes resources
via the `*` selector.
This is not needed, because the policy knows exactly what kind of resources have
been deprecated, hence it can listen only for events related to these resource types.

Why we needed to change this behaviour:

* Reduce the load on the policy server: listen only for relevant events
* Make it easy for the audit scan to inspect this policy

Fixes https://github.com/kubewarden/deprecated-api-versions-policy/issues/4

Signed-off-by: Flavio Castelli <fcastelli@suse.com>