fix!: change the tagging strategy
Do not use `-` when building the tag of our releases, this causes the
helm chart of this policy to be seen as a development release, hence
it's hidden from the standard result.

We cannot use `+` instead, because that would result in an invalid OCI
url (`+` symbols are not allowed by container image references).

Because of that, we've dropped the kubernetes annotation from the policy
tag and moved that into the README.

Finally, more sanity checks have been added to the GHA.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>